Move oracle/roll state code to own file
